Change components locations and map state and props directly on components
