+++ src/components/App.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import BookList from '../containers/BookList';
+import BooksForm from '../containers/BooksForm';
+
+const App = () => (
+  <div className="App">
+    <BookList />
+    <BooksForm />
+  </div>
+);
+
+export default App;
