+++ src/index.js
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
 import { createStore } from 'redux';
 import './index.css';
-import App from './containers/App';
+import App from './components/App';
 import rootReducer from './reducers';
 
 const store = createStore(rootReducer);
